Implement proper API version structure
Converts v1 to ES6
