+++ docker_node_server/app.js
@@ -11,7 +11,6 @@
   Modules
 */
 import express from 'express';
-import bodyParser from 'body-parser';
 import cors from 'cors';
 import helmet from 'helmet';
 import morgan from 'morgan';
@@ -23,21 +22,14 @@ const app = express();
 app.use(helmet());
 app.use(cors());
 app.use(morgan('combined'));
-app.use(bodyParser.urlencoded({ extended: false }));
-app.use(bodyParser.json());
+app.use(express.urlencoded({ extended: false }));
+app.use(express.json());
 
-/*
-  V1 Routes
-  imports of the routes from the v1 directory
-*/
-import {router as secret_submit_route } from './v1/submit_secret.js';
-app.use(secret_submit_route);
+import { router as v1 } from './src/v1/index.js'
+import { router as v2 } from './src/v2/index.js'
 
-import {router as secret_get_route } from './v1/get_secret.js';
-app.use(secret_get_route);
-
-import {router as info_for_nerds_route } from './v1/info_for_nerds.js';
-app.use(info_for_nerds_route);
+app.use('/api/v1', v1)
+app.use('/api/v2', v2)
 
 let server = app.listen(process.env.PORT || 3001, () => {
   console.log('listening on port ' + server.address().port)
