+++ docker_node_server/package.json
@@ -19,6 +19,7 @@
     "express": "^4.18.1",
     "helmet": "^5.0.2",
     "morgan": "^1.10.0",
-    "nodemon": "^2.0.16"
+    "nodemon": "^2.0.16",
+    "sqlite3": "^5.0.8"
   }
 }
