--- docker_node_server/v0.1/index.js
@@ -1,17 +1,0 @@
-var router = require('express').Router()
-var { import_test } = require('./utilities/test')
-var db = require('../../services/db')
-
-db.initialise()
-
-router.use('/secret', require('./secret'))
-router.use('/passphrase', require('./passphrase'))
-router.use('/test', require('./crypto_test'))
-
-router.get('/', (req, res) => {
-  res.status(200).send({
-    '_': import_test()
-  })
-})
-
-module.exports = router
